[new] - lol_spell_recognition.py file created.
With creation, coordinates of all players' spells in the image are written in comments.

Top-left to bottom-right coordinates added to the file as comments.
The coordinates should be used in the order written in comments to access the correct pixel.
In other words, the comments write y-coordinate and x-coordinate respectively.